# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PortalHatal Balmas API",
        "version": "2"
    },
    "paths": {
...
        "/api/info": {
            "get": {
                "tags": [
                    "Informations"
                ],
                "summary": "מקבל את רשימת הכתבות מידע",
                "description": "מחזיר רשימה של כל הכתבות מידע במערכת.",
                "operationId": "getInformations",
                "responses": {
                    "200": {
                        "description": "הפעולה בוצעה בהצלחה"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            },
            "post": {
                "tags": [
                    "Informations"
                ],
                "summary": "יוצר כתבת מידע חדש",
                "description": "יוצר כתבת מידע חדש עם כותרת, תוכן ותמונה",
                "operationId": "createInformation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "כותרת הכתבת מידע"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "{blah:{dfgfdgfd}"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "הכתבת מידע נוצר בהצלחה"
                    },
                    "400": {
                        "description": "נתונים חסרים או שגויים"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            }
        },
        "/api/info/{uuid}": {
            "get": {
                "tags": [
                    "Informations"
                ],
                "summary": "מביא פוסט לפי UUID",
                "description": "מביא פוסט בודד לפי מזהה ייחודי (UUID).",
                "operationId": "getInformationByUUid",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID של הכרזה",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "4a206b4-99d6-4692-915c-4935766e0420"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה בוצעה בהצלחה"
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "404": {
                        "description": "הכרזה לא נמצא"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            },
            "post": {
                "tags": [
                    "Informations"
                ],
                "summary": "מעדכן פוסט ",
                "description": "מעדכן פוסט .",
                "operationId": "updateInformation",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID של כתבת מידע",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "4a206b4-99d6-4692-915c-4935766e0420"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "כותרת כתבת מידע"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "תיאור קצר של כתבת מידע"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "{blah:{dfgfdgfd}"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "description": "תמונה מצורפת לפוסט",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "כתבת מידע נוצר בהצלחה"
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Informations"
                ],
                "summary": "מוחק כתבת מידע לפי UUID",
                "description": "מוחק כתבת מידע בודד לפי מזהה ייחודי (UUID).",
                "operationId": "deleteInformationByUUid",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID של הכתבת מידע",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "4a206b4-99d6-4692-915c-4935766e0420"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה בוצעה בהצלחה"
                    },
                    "404": {
                        "description": "הכתבת מידע לא נמצא"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}